+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version [7.0.0](https://github.com/cedx/coveralls.js/compare/v6.0.0...v7.0.0)
 - Breaking change: ported the source code to [TypeScript](https://www.typescriptlang.org).
 - Breaking change: removed the `defaultValue` argument from the `Configuration.get()` method.
+- Breaking change: the `Configuration.fromEnvironment` is now asynchronous.
 - Added the `eventRequest` and `eventResponse` static properties to the `Client` class.
 - Replaced [ESDoc](https://esdoc.org) documentation generator by [TypeDoc](https://typedoc.org).
 - Updated the package dependencies.